requirements.txt
amqp==5.0.9
APScheduler==3.8.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.2
django-apscheduler==0.6.0
django-filter==21.1
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.1.0
openapi-codec==1.3.2
packaging==21.3
phonenumberslite==8.12.43
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
